add test to test all the username fragment and correct the bug that don't take mockdatabase during the test phases
