+++ app/src/main/java/com/github/drawyourpath/bootcamp/database/MockDataBase.kt
@@ -1,7 +1,6 @@
 package com.github.drawyourpath.bootcamp.database
 
 import android.graphics.Color
-import android.widget.EditText
 import android.widget.TextView
 
 class MockDataBase : Database() {
@@ -11,24 +10,24 @@ class MockDataBase : Database() {
         //add an element in the user database to test the UI
         usersDataBase.add("albert")
         val unAvailableOutput: String = buildString {
-            append("*The userID ")
+            append("*The username ")
             append(userName)
             append(" is NOT available !")
         }
         val availableOutput: String = buildString {
-            append("*The userID ")
+            append("*The username ")
             append(userName)
             append(" is available !")
         }
         if (userName == "") {
-            outputText.text = buildString { append("The user name can't be empty !") }
+            outputText.text = buildString { append("The username can't be empty !") }
             outputText.setTextColor(Color.RED)
             return false
-        }else if(usersDataBase.contains(userName)){
+        } else if (usersDataBase.contains(userName)) {
             outputText.text = unAvailableOutput
             outputText.setTextColor(Color.RED)
             return false
-        }else {
+        } else {
             outputText.text = availableOutput
             outputText.setTextColor(Color.GREEN)
             return true
@@ -36,7 +35,7 @@ class MockDataBase : Database() {
     }
 
     override fun setUserName(userName: String, outputText: TextView) {
-        if(isUserNameAvailable(userName, outputText)){
+        if (isUserNameAvailable(userName, outputText)) {
             usersDataBase.add(userName)
         }
     }
